ONE SCRIPT TO RULE THEM ALL AND IN THE DARKNESS BIND THEM
